+++ Core/Source/Renderer/Rectangle.h
@@ -1,12 +1,13 @@
 #pragma once
+#include <cstdint>
 
 namespace Mupfel {
 
 	class Rectangle
 	{
 	public:
-		static void RaylibDrawRect(int posX, int posY, int width, int height, int r, int g, int b, int alpha);
-		static void RaylibDrawRectFilled(int posX, int posY, int width, int height, int r, int g, int b, int alpha);
+		static void RaylibDrawRect(int posX, int posY, int width, int height, uint8_t r, uint8_t g, uint8_t b, uint8_t alpha);
+		static void RaylibDrawRectFilled(int posX, int posY, int width, int height, uint8_t r, uint8_t g, uint8_t b, uint8_t alpha);
 	};
 
 }
